actions: Major worker refactoring
Change-type: patch
