+++ lib/actions/errors.js
@@ -19,6 +19,7 @@ const typedErrors = require('typed-errors')
 
 _.each([
 	'ActionsNoElement',
+	'ActionsSchemaMismatch',
 	'ActionsAuthenticationError'
 ], (error) => {
 	exports[error] = typedErrors.makeTypedError(error)
