+++ lib/actions/index.js
@@ -22,9 +22,86 @@ const objectTemplate = require('object-template')
 const skhema = require('skhema')
 const library = require('./library')
 const time = require('./time')
-const Orchestrator = require('./orchestrator')
 const utils = require('./utils')
+const errors = require('./errors')
 const jellyscript = require('../jellyscript')
+
+const TYPE_ACTION_REQUEST = 'action-request'
+
+const SCHEMA_ACTIVE_TRIGGERS = {
+	type: 'object',
+	properties: {
+		active: {
+			type: 'boolean',
+			const: true
+		},
+		type: {
+			type: 'string',
+			const: 'triggered-action'
+		},
+		data: {
+			type: 'object',
+			additionalProperties: true
+		}
+	},
+	required: [ 'active', 'type', 'data' ]
+}
+
+const SCHEMA_PENDING_REQUESTS = {
+	type: 'object',
+	required: [ 'id', 'active', 'type', 'data' ],
+	properties: {
+		id: {
+			type: 'string'
+		},
+		active: {
+			type: 'boolean',
+			const: true
+		},
+		type: {
+			type: 'string',
+			const: TYPE_ACTION_REQUEST
+		},
+		data: {
+			type: 'object',
+			required: [ 'action', 'actor', 'target', 'executed', 'arguments' ],
+			properties: {
+				action: {
+					type: 'string'
+				},
+				actor: {
+					type: 'string'
+				},
+				target: {
+					type: 'string'
+				},
+				executed: {
+					type: 'boolean',
+					const: false
+				},
+				arguments: {
+					type: 'object',
+					additionalProperties: true
+				}
+			}
+		}
+	}
+}
+
+const closeStream = (streams, name) => {
+	if (!streams[name]) {
+		return Bluebird.resolve()
+	}
+
+	return new Bluebird((resolve) => {
+		streams[name].once('closed', () => {
+			Reflect.deleteProperty(streams, name)
+			resolve()
+		})
+
+		streams[name].close()
+	})
+}
 
 const slugify = (string) => {
 	return string
@@ -32,12 +109,106 @@ const slugify = (string) => {
 		.replace(/-{1,}/g, '-')
 }
 
+const processRequest = (worker, session, request) => {
+	const data = request.data
+
+	return Bluebird.resolve(worker.executeAction(session, {
+		actionId: data.action,
+		targetId: data.target,
+		actorId: data.actor
+	}, data.arguments)).then((results) => {
+		debug(`Action request executed: ${request.id}`)
+
+		return worker.executeAction(session, {
+			actionId: 'action-update-card',
+			targetId: request.id,
+			actorId: data.actor
+		}, {
+			properties: {
+				data: {
+					executed: true,
+					result: {
+						error: false,
+						timestamp: time.getCurrentTimestamp(),
+						data: results
+					}
+				}
+			}
+		})
+	}).catch((error) => {
+		return worker.executeAction(session, {
+			actionId: 'action-update-card',
+			targetId: request.id,
+			actorId: data.actor
+		}, {
+			properties: {
+				data: {
+					executed: true,
+					result: {
+						error: true,
+						timestamp: time.getCurrentTimestamp(),
+						data: {
+							type: error.name,
+							message: error.message
+						}
+					}
+				}
+			}
+		})
+	}).tap(() => {
+		debug(`Action request results set: ${request.id}`)
+	})
+}
+
+const refreshTypeCards = async (jellyfish, worker) => {
+	worker.typeCards.card = await jellyfish.getCardBySlug(worker.session, 'card', {
+		type: 'type'
+	})
+
+	if (!worker.typeCards.card) {
+		throw new Error('There is no card type: card')
+	}
+
+	worker.typeCards.action = await jellyfish.getCardBySlug(worker.session, 'action', {
+		type: 'type'
+	})
+
+	if (!worker.typeCards.action) {
+		throw new Error('There is no card type: action')
+	}
+}
+
+const getActorKey = async (jellyfish, session, actorId) => {
+	const keySlug = `session-action-${actorId}`
+	const key = await jellyfish.getCardBySlug(session, keySlug, {
+		type: 'session'
+	})
+
+	if (key && key.data.actor === actorId) {
+		return key
+	}
+
+	return jellyfish.insertCard(session, {
+		slug: keySlug,
+		type: 'session',
+		active: true,
+		links: [],
+		tags: [],
+		data: {
+			actor: actorId
+		}
+	}, {
+		override: true
+	})
+}
+
 class ActionRequestWorker {
 	constructor (jellyfish, session) {
-		this.orchestrator = new Orchestrator(jellyfish, session)
 		this.jellyfish = jellyfish
 		this.session = session
 		this.typeCards = {}
+		this.triggers = []
+		this.streams = {}
 	}
 
 	async getCardBySlug (session, slug) {
@@ -77,7 +248,7 @@ class ActionRequestWorker {
 
 		const typeSchema = _.get(typeCard, [ 'data', 'schema' ], null)
 		if (!typeSchema) {
-			throw new this.jellyfish.errors.JellyfishUnknownCardType(`Unknown type: ${properties.type}`)
+			throw new errors.ActionsNoElement(`Unknown type: ${properties.type}`)
 		}
 
 		const result = jellyscript.evaluateObject(typeSchema, properties)
@@ -161,7 +332,7 @@ class ActionRequestWorker {
 				})
 
 				if (!triggeredActionTypeCard) {
-					throw new this.jellyfish.errors.JellyfishNoElement(`No such type: ${triggeredActionSlug}`)
+					throw new errors.ActionsNoElement(`No such type: ${triggeredActionSlug}`)
 				}
 
 				const targetProperty = _.join(_.concat([ 'source' ], watcher.target), '.')
@@ -192,64 +363,45 @@ class ActionRequestWorker {
 		return insertedCard
 	}
 
-	async refreshTypeCards () {
-		this.typeCards.card = await this.jellyfish.getCardBySlug(this.session, 'card', {
-			type: 'type'
-		})
-
-		if (!this.typeCards.card) {
-			throw new Error('There is no card type: card')
-		}
-
-		this.typeCards.action = await this.jellyfish.getCardBySlug(this.session, 'action', {
-			type: 'type'
-		})
-
-		if (!this.typeCards.action) {
-			throw new Error('There is no card type: action')
-		}
-	}
-
 	async refreshTriggers () {
-		return this.orchestrator.refreshTriggers()
-	}
-
-	async getActorKey (actorId) {
-		const keySlug = `session-action-${actorId}`
-		const key = await this.jellyfish.getCardBySlug(this.session, keySlug, {
-			type: 'session'
-		})
-
-		if (key && key.data.actor === actorId) {
-			return key
-		}
-
-		return this.jellyfish.insertCard(this.session, {
-			slug: keySlug,
-			type: 'session',
-			active: true,
-			links: [],
-			tags: [],
-			data: {
-				actor: actorId
-			}
-		}, {
-			override: true
-		})
+		this.triggers =
+			await this.jellyfish.query(this.session, SCHEMA_ACTIVE_TRIGGERS)
 	}
 
 	async start (options = {}) {
-		const emitter = new EventEmitter()
-		this.refreshTypeCards()
+		await this.stop()
+		await this.refreshTriggers()
+		await refreshTypeCards(this.jellyfish, this)
 
-		this.orchestrator.on('error', (error) => {
+		const emitter = new EventEmitter()
+
+		this.streams.triggers =
+			await this.jellyfish.stream(this.session, SCHEMA_ACTIVE_TRIGGERS)
+		this.streams.triggers.on('data', this.refreshTriggers.bind(this))
+		this.streams.triggers.on('error', (error) => {
 			emitter.emit('error', error)
 		})
 
-		this.orchestrator.on('request', (request) => {
+		this.streams.pendingRequests =
+			await this.jellyfish.stream(this.session, SCHEMA_PENDING_REQUESTS)
+		this.streams.pendingRequests.on('error', (error) => {
+			emitter.emit('error', error)
+		})
+
+		this.streams.pendingRequests.on('data', (change) => {
+			emitter.emit('pending-request', change.after)
+		})
+
+		Bluebird.resolve(this.getPendingRequests()).each((request) => {
+			emitter.emit('pending-request', request)
+		}).catch((error) => {
+			emitter.emit('error', error)
+		})
+
+		emitter.on('pending-request', (request) => {
 			debug(`Incoming action request ${request.id} (${request.data.action})`)
-			return this.getActorKey(request.data.actor).then((key) => {
-				return this.processRequest(key.id, request).then(() => {
+			return getActorKey(this.jellyfish, this.session, request.data.actor).then((key) => {
+				return processRequest(this, key.id, request).then(() => {
 					emitter.emit('request', request)
 				})
 			}).catch((error) => {
@@ -257,13 +409,15 @@ class ActionRequestWorker {
 			})
 		})
 
-		await this.orchestrator.start()
 		return emitter
 	}
 
 	async stop () {
 		await this.flushPendingRequests()
-		return this.orchestrator.stop()
+		return Bluebird.all([
+			closeStream(this.streams, 'triggers'),
+			closeStream(this.streams, 'pendingRequests')
+		])
 	}
 
 	async flushPendingRequests () {
@@ -273,73 +427,53 @@ class ActionRequestWorker {
 		}
 
 		for (const request of pendingRequests) {
-			const key = await this.getActorKey(request.data.actor)
-			await this.processRequest(key.id, request)
+			const key = await getActorKey(this.jellyfish, this.session, request.data.actor)
+			await processRequest(this, key.id, request)
 		}
 
 		return this.flushPendingRequests()
 	}
 
 	async getPendingRequests () {
-		return this.orchestrator.getPendingRequests()
-	}
-
-	async processRequest (session, request) {
-		const data = request.data
-
-		return Bluebird.resolve(this.executeAction(session, {
-			actionId: data.action,
-			targetId: data.target,
-			actorId: data.actor
-		}, data.arguments)).then((results) => {
-			debug(`Action request executed: ${request.id}`)
-
-			return this.executeAction(session, {
-				actionId: 'action-update-card',
-				targetId: request.id,
-				actorId: data.actor
-			}, {
-				properties: {
-					data: {
-						executed: true,
-						result: {
-							error: false,
-							timestamp: time.getCurrentTimestamp(),
-							data: results
-						}
-					}
-				}
-			})
-		}).catch((error) => {
-			return this.executeAction(session, {
-				actionId: 'action-update-card',
-				targetId: request.id,
-				actorId: data.actor
-			}, {
-				properties: {
-					data: {
-						executed: true,
-						result: {
-							error: true,
-							timestamp: time.getCurrentTimestamp(),
-							data: {
-								type: error.name,
-								message: error.message
-							}
-						}
-					}
-				}
-			})
-		}).tap(() => {
-			debug(`Action request results set: ${request.id}`)
-		})
+		return this.jellyfish.query(this.session, SCHEMA_PENDING_REQUESTS)
 	}
 
 	async createRequest (session, options) {
-		return this.orchestrator.createRequest(session, options.action, {
-			targetId: options.targetId,
-			actorId: options.actorId,
-			arguments: options.arguments
+		const actionCard = await this.jellyfish.getCardBySlug(session, options.action, {
+			type: 'action',
+			writeMode: true
+		})
+
+		if (!actionCard) {
+			throw new Error(`You don't have permission to execute this action: ${options.action}`)
+		}
+
+		return this.jellyfish.insertCard(session, {
+			type: TYPE_ACTION_REQUEST,
+			active: true,
+			links: [],
+			tags: [],
+			data: {
+				action: actionCard.slug,
+
+				// Anyone is free to create an action request
+				// matching to anything they like. Whether they
+				// can access these cards is something that must
+				// be checked right before actually executing the
+				// action, otherwise if all the checks are here,
+				// then someone can circunvent them all by creating
+				// a request directly on the core server, which
+				// won't perform any checks.
+				actor: options.actorId,
+				target: options.targetId,
+
+				timestamp: time.getCurrentTimestamp(),
+				executed: false,
+				arguments: jellyscript.evaluateObject(
+					utils.getActionArgumentsSchema(actionCard),
+					options.arguments
+				).object
+			}
 		})
 	}
 
@@ -355,10 +489,10 @@ class ActionRequestWorker {
 		})
 
 		if (!sessionCard) {
-			throw new this.jellyfish.errors.JellyfishNoElement('Invalid session')
+			throw new errors.ActionsNoElement('Invalid session')
 		}
 
-		const requests = await Bluebird.map(this.orchestrator.getTriggeredActionCards(), (trigger) => {
+		const requests = await Bluebird.map(this.triggers, (trigger) => {
 			const compiledTrigger = _.attempt(objectTemplate.compile, trigger.data, {
 				source: card
 			}, {
@@ -389,7 +523,7 @@ class ActionRequestWorker {
 		})
 
 		if (!actionCard) {
-			throw new this.jellyfish.errors.JellyfishNoAction(`Unknown action: ${options.actionId}`)
+			throw new errors.ActionsNoElement(`Unknown action: ${options.actionId}`)
 		}
 
 		let targetCard = await this.jellyfish.getCardById(session, options.targetId)
@@ -397,7 +531,7 @@ class ActionRequestWorker {
 			targetCard = await this.jellyfish.getCardBySlug(session, options.targetId)
 		}
 		if (!targetCard) {
-			throw new this.jellyfish.errors.JellyfishNoElement(`Unknown target: ${options.targetId}`)
+			throw new errors.ActionsNoElement(`Unknown target: ${options.targetId}`)
 		}
 
 		const actor = await this.jellyfish.getCardById(session, options.actorId)
@@ -406,26 +540,26 @@ class ActionRequestWorker {
 		}
 
 		if (!this.typeCards.card || !this.typeCards.action) {
-			await this.refreshTypeCards()
+			await refreshTypeCards(this.jellyfish, this)
 		}
 
 		if (!skhema.isValid(this.typeCards.card.data.schema, actionCard)) {
-			throw new this.jellyfish.errors.JellyfishSchemaMismatch(`Invalid action: ${actionCard.id}`)
+			throw new errors.ActionsSchemaMismatch(`Invalid action: ${actionCard.id}`)
 		}
 
 		if (!skhema.isValid(this.typeCards.action.data.schema, actionCard)) {
-			throw new this.jellyfish.errors.JellyfishSchemaMismatch(`Invalid action: ${actionCard.id}`)
+			throw new errors.ActionsSchemaMismatch(`Invalid action: ${actionCard.id}`)
 		}
 
 		if (!skhema.isValid(_.get(actionCard, [ 'data', 'filter' ], {
 			type: 'object'
 		}), targetCard)) {
-			throw new this.jellyfish.errors.JellyfishSchemaMismatch('Target does not match filter')
+			throw new errors.ActionsSchemaMismatch('Target does not match filter')
 		}
 
 		const argumentsSchema = utils.getActionArgumentsSchema(actionCard)
 		if (!skhema.isValid(argumentsSchema, args)) {
-			throw new this.jellyfish.errors.JellyfishSchemaMismatch('Arguments do not match')
+			throw new errors.ActionsSchemaMismatch('Arguments do not match')
 		}
 
 		const compiledArguments = objectTemplate.compile(actionCard.data.options, {
@@ -448,11 +582,11 @@ class ActionRequestWorker {
 
 		const actionFunction = library[actionCard.slug]
 		if (!actionFunction) {
-			throw new this.jellyfish.errors.JellyfishNoAction(`Unknown action function: ${actionCard.slug}`)
+			throw new errors.ActionsNoElement(`Unknown action function: ${actionCard.slug}`)
 		}
 
 		if (!await this.jellyfish.getCardById(session, options.actorId)) {
-			throw new this.jellyfish.errors.JellyfishNoElement(`Unknown actor: ${options.actorId}`)
+			throw new errors.ActionsNoElement(`Unknown actor: ${options.actorId}`)
 		}
 
 		debug(`Executing internal action ${actionCard.slug}`)
@@ -473,7 +607,7 @@ module.exports.setup = async (jellyfish, session, options) => {
 	})
 
 	if (!sessionCard) {
-		throw new jellyfish.errors.JellyfishNoElement('Invalid session')
+		throw new errors.ActionsNoElement('Invalid session')
 	}
 
 	const adminWorker = new ActionRequestWorker(jellyfish, session)
@@ -482,7 +616,7 @@ module.exports.setup = async (jellyfish, session, options) => {
 	})
 
 	if (!userCard) {
-		throw new jellyfish.errors.JellyfishNoElement('The user card does not exist')
+		throw new errors.ActionsNoElement('The user card does not exist')
 	}
 
 	if (!await jellyfish.getCardBySlug(session, `user-${options.username}`, {
@@ -503,7 +637,7 @@ module.exports.setup = async (jellyfish, session, options) => {
 			}
 		})
 
-		await adminWorker.processRequest(session, signupRequest)
+		await processRequest(adminWorker, session, signupRequest)
 	}
 
 	const actor = await jellyfish.getCardBySlug(session, `user-${options.username}`, {
@@ -512,7 +646,7 @@ module.exports.setup = async (jellyfish, session, options) => {
 	})
 
 	if (!actor) {
-		throw new jellyfish.errors.JellyfishNoElement(`Unknown user: ${options.username}`)
+		throw new errors.ActionsNoElement(`Unknown user: ${options.username}`)
 	}
 
 	// Set the role for the actions user, as it requires a high level of privilege
@@ -530,7 +664,7 @@ module.exports.setup = async (jellyfish, session, options) => {
 			}
 		})
 
-		await adminWorker.processRequest(session, updateRequest)
+		await processRequest(adminWorker, session, updateRequest)
 	}
 
 	const loginRequest = await adminWorker.createRequest(session, {
@@ -547,6 +681,6 @@ module.exports.setup = async (jellyfish, session, options) => {
 		}
 	})
 
-	const completedLoginRequest = await adminWorker.processRequest(session, loginRequest)
+	const completedLoginRequest = await processRequest(adminWorker, session, loginRequest)
 	return completedLoginRequest.data.result.data
 }
